Add Maven FB

+++ fb/FBManager.sh
@@ -3,6 +3,7 @@
 # Arguments
 # 1: Configuration file JSON Format See help description
 # 2: Action: create / delete Feature Branch
+# 3: Feature Branch Name: word "feature/" will be added automatically in case of unset
 
 # EXIT CODES
 # 0 -- OK
@@ -16,13 +17,14 @@ protected_branches_list_regex="^(master|develop|stable\/[0-9.-_a-zA-Z]+)$"
 ####
 
 scrdir=$(dirname $(realpath $0))
-source ${scrdir}/_functions.sh
+. ${scrdir}/_functions.sh
 if [ -z "$1" ] || [[ $1 =~ ^-(h|-help)$ ]]; then
     print_help
     exit 0
 fi
 assert_command git
 assert_command jq
+assert_command mvn
 if [ -z "$1" ] || [ ! -f "$1" ]; then
     echo_err "Config file is missing or invalid"
     exit 1
@@ -41,9 +43,25 @@ if [[ ! "$1" =~ ^(create|delete)$ ]]; then
 else
     echo_ok "Action \"$1\" will be performed"
 fi
-
 action=$1
-start_time=$(date +%s);
+shift
+if [ -z "$1" ]; then
+    echo_err "Missing Feature Branch name!!"
+    exit 2
+fi
+featurebranch=$1
+[[ "${featurebranch}" =~ ^Feature ]] || featurebranch="feature/${featurebranch}"
+
+maventag=$(echo ${featurebranch} | sed -E 's|^(f\|F)eature/||g')
+#Check for illegal characteres like / etc 
+if [[ ! "${maventag}" =~ ^[A-Za-z0-9_-]+$ ]]; then
+    echo_err "Feature branch \"${featurebranch}\" contains illegal characters!"
+    exit 3
+fi
+
+shift
+echo_ok "Feature Branch is \"${featurebranch}\""
+start_time=$(date +%s)
 rm -rf ${wkdir}* &>/dev/null #Cleanup
 for el in $(cat ${config_file} | jq -c '.[]'); do
     ## Args Parsing
@@ -61,16 +79,6 @@ for el in $(cat ${config_file} | jq -c '.[]'); do
 
     if [ -z "${git_repository}" ]; then
         echo_err "Github repository is not specified!"
-        exit 3
-    fi
-
-    if [ -z "${name}" ]; then
-        echo_err "Github target branch is not specified!"
-        exit 3
-    fi
-
-    if [[ "${name}" =~ ${protected_branches_list_regex} ]]; then
-        echo_err "Github target branch \"${name}\" matches a protected branch"
         exit 3
     fi
 
@@ -103,7 +111,7 @@ for el in $(cat ${config_file} | jq -c '.[]'); do
 
     ## Function Live Definition Make code below more redeable
     r_git() {
-        git --git-dir=${local_repo_path}/.git $*
+        git --git-dir=${local_repo_path}/.git --work-tree=${local_repo_path} $*
     }
 
     ## Check if update is unset if yes, use false as default value
@@ -115,10 +123,10 @@ for el in $(cat ${config_file} | jq -c '.[]'); do
         ## Check if base branch is unset if yes, use default branch as base one
         [ -z "${git_base_branch}" ] && git_base_branch=$(r_git symbolic-ref --short HEAD)
 
-        ## Check if target branch is already exist in the Git Repository [ Remote Check ]
-        if [ ! -z "$(r_git ls-remote --heads origin ${name})" ]; then
+        ## Check if feature branch is already exist in the Git Repository [ Remote Check ]
+        if [ ! -z "$(r_git ls-remote --heads origin ${featurebranch})" ]; then
             if [[ ${update} == "false" ]]; then
-                echo_err "Branch \"${name}\" already exist!"
+                echo_err "Branch \"${featurebranch}\" already exist!"
             fi
             if [[ ${update} == "true" ]]; then
                 echo_warn "Update Mode is Enabled!"
@@ -129,8 +137,8 @@ for el in $(cat ${config_file} | jq -c '.[]'); do
 
         current_branch=$(r_git rev-parse --abbrev-ref HEAD)
         #Default branch protection
-        if [[ "${name}" == "${current_branch}" ]]; then
-            echo_err "Could not recreate the default branch \"${name}\"!"
+        if [[ "${featurebranch}" == "${current_branch}" ]]; then
+            echo_err "Could not recreate the default branch \"${featurebranch}\"!"
             exit_with_cleanup 4
         fi
         ## Check if the current branch is the selected base branch if not, check out to it
@@ -147,36 +155,68 @@ for el in $(cat ${config_file} | jq -c '.[]'); do
         # Add Force flag for the git push and remove branch as precaution
         force_flag=""
         [[ ${update} == "true" ]] && force_flag="-f"
-        r_git branch -D ${name} &>/dev/null
+        r_git branch -D ${featurebranch} &>/dev/null
 
-        # Create target branch locally and push to the remote
-        if r_git checkout -b "${name}" && r_git push origin "${name}" ${force_flag} 2>/dev/null; then
-            echo_ok "Branch \"${name}\" has been created!"
+        # Create feature branch locally and push to the remote
+        if r_git checkout -b "${featurebranch}" 2>/dev/null; then
+            echo_ok "Branch \"${featurebranch}\" has been created locally"
         else
-            echo_err "Could not create branch \"${name}\"!"
+            echo_err "Could not create branch \"${featurebranch}\"!"
             exit_with_cleanup 5
+        fi
+
+        # Apply changes to Maven Project
+        if [ -f "${local_repo_path}/pom.xml" ]; then
+            project_artifactId=$(get_maven_property "${local_repo_path}" "project.artifactId")
+            project_groupId=$(get_maven_property "${local_repo_path}" "project.groupId")
+            project_version=$(get_maven_property "${local_repo_path}" "project.version")
+            new_project_version=$(echo ${project_version} | sed "s|-SNAPSHOT|-${maventag}-SNAPSHOT|g")
+            echo_ok "Maven Project Informations:"
+            echo "***"
+            echo "       ArtifactId: ${project_artifactId}"
+            echo "          GroupId: ${project_groupId}"
+            echo "          Version: ${project_version}"
+            echo "      New version: ${new_project_version}"
+            echo "***"
+            if change_maven_project_version "${local_repo_path}" ${project_groupId} ${project_artifactId} ${new_project_version}; then
+                echo_ok "Maven Project version changed to ${new_project_version}"
+            else
+                echo_err "Could not change project version to ${new_project_version}"
+                exit_with_cleanup 4
+            fi
+            r_git add $(find ${local_repo_path} -name pom.xml | xargs)
+            if ! echo "Create Feature Branch ${featurebranch}" | r_git commit -F -; then
+                echo_err "Could not commit on the ${featurebranch} branch!"
+                exit_with_cleanup 4
+            fi
+        fi
+
+        if ! r_git push origin "${featurebranch}" ${force_flag} 2>/dev/null; then
+            echo_err "Could not push on the ${featurebranch} branch!"
+            exit_with_cleanup 4
+        else
+            echo_ok "Branch ${featurebranch} has been created"
         fi
 
     ## Perform Feature Branch Deletion
     elif [[ $action == "delete" ]]; then
         # Default Branch Protection
         current_branch=$(r_git rev-parse --abbrev-ref HEAD)
-        if [[ "${name}" == "${current_branch}" ]]; then
-            echo_err "Could not delete default branch \"${name}\"!"
+        if [[ "${featurebranch}" == "${current_branch}" ]]; then
+            echo_err "Could not delete default branch \"${featurebranch}\"!"
             exit_with_cleanup 4
         fi
-        ## Check if target branch is already exist in the Git Repository [ Remote Check ]
-        if [ ! -z "$(r_git ls-remote --heads origin ${name})" ]; then
-            echo_ok "Branch \"${name}\" exist"
-            if r_git push origin :${name}; then
-                echo_ok "Branch \"${name}\" has been deleted!"
-
+        ## Check if feature branch is already exist in the Git Repository [ Remote Check ]
+        if [ ! -z "$(r_git ls-remote --heads origin ${featurebranch})" ]; then
+            echo_ok "Branch \"${featurebranch}\" exist"
+            if r_git push origin :${featurebranch}; then
+                echo_ok "Branch \"${featurebranch}\" has been deleted!"
             else
-                echo_err "Could not delete branch \"${name}\"!"
+                echo_err "Could not delete branch \"${featurebranch}\"!"
                 exit_with_cleanup 5
             fi
         else
-            echo_ok "Branch \"${name}\" does not exist"
+            echo_ok "Branch \"${featurebranch}\" does not exist"
         fi
     fi
     ## Mandatory Common cleanup process for the next iteration
